Adding the process.php page / controller
This page will allow a user to register or log in to the application
system. It will query the database and verify that the credentials
passed to it (if logging in) match a pair in the database, then set some
sessions and log the user in.

Otherwise, if they are registering, it will try to add the credentials
to the database to create a new user 'by saving their username, password
and salt'.

At the moment, the scripts are far from perfect, and still need to
adjusted to account for duplication and sql injection and to hold
various other fields. I will work on getting this stuff down and working
after my exam tomorrow; unless someone just totally feels up to taking
it on.